requirements.txt
joblib==1.3.2
pandas==2.1.4
Pillow==10.1.0
streamlit==1.29.0
